+++ src/noise/noise_model.rs
@@ -1,4 +1,4 @@
 #[derive(Clone, Copy, Debug)]
 pub enum NoiseType {
-    Depolarizing(f32),
+    Depolarizing(f64),
 }
